benchmarks/requirements.txt
langchain-text-splitters>=0.3.0
chonkie[all]>=1.6.0
llama-index-core>=0.12.0
haystack-ai>=2.8.0
psutil>=5.9.0
